➡️ Refactored WhisperTabCompleter Moved it into TabCompleters Package
